requirements.txt
streamlit>=1.32.0
pydantic>=2.6.0
pandas>=2.2.0
plotly>=5.18.0
nltk==3.8.1
python-dotenv>=1.0.0
anthropic>=0.18.0
networkx>=3.2.0
transformers>=4.38.0
torch>=2.2.0
numpy>=1.24.0
spacy>=3.7.0
textblob>=0.17.1
tabulate>=0.9.0 